requirements.txt
asgiref==3.6.0
coverage==7.0.5
Django==4.1.5
django-rest-framework==0.1.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
Pillow==9.4.0
psycopg2==2.9.5
python-dotenv==0.21.0
pytz==2022.7.1
sqlparse==0.4.3
